chore(refactor): Refactor code for custom orders route (#6610)
