+++ app/__init__.py
@@ -146,11 +146,12 @@ def create_app():
         from app.api.users import user_misc_routes
         from app.api.orders import order_misc_routes
         from app.api.role_invites import role_invites_misc_routes
-        from app.api.auth import ticket_blueprint, authorised_blueprint
+        from app.api.auth import authorised_blueprint
         from app.api.admin_translations import admin_blueprint
         from app.api.orders import alipay_blueprint
         from app.api.settings import admin_misc_routes
         from app.api.server_version import info_route
+        from app.api.custom.orders import ticket_blueprint
 
         app.register_blueprint(api_v1)
         app.register_blueprint(event_copy)
